l3rad/requirements.txt
flask
flask-caching
google-cloud-storage
gunicorn
pillow
numpy
